Create a wrapper script which looks for an executable in the workspace
This executable in the workspace can be another Bazel binary whose
version will change with the code it's next to, or a shell script which
downloads a fixed version from some location.

RELNOTES: A tools/bazel script in the workspace will be executed
as an opportunity to use a fixed version of Bazel (not implemented for
the homebrew recipe yet).

Fixes #521

--
Change-Id: Id06177d9c2b259cd9d6fd62edb5abe541342dd05
Reviewed-on: https://bazel-review.googlesource.com/2620
MOS_MIGRATED_REVID=112477232