Fix documentation of --[no]block_for_lock
The help message of this option had the effect inversed. The default
--block_for_lock setting makes the client block waiting for the lock
when needed, and the flag needs to be unset with --noblock_for_lock
to instead exit immediately.

Also rename the dummy noblock_for_lock variable to blockForLock to match
the naming convention. The variable is not used anywhere, as this class
only exists for the help messages.

--
MOS_MIGRATED_REVID=112566287