Support @ syntax for specifying the list of files to zipper
When creating an archive with //third_party/ijar:zipper, all
files to add should be specified as an argument. When adding a
lot of files, this leads to reaching the maximum size of the
command line.

This change add a @filename argument that permit to specify a
file containing the list of files to zip.

--
Change-Id: I02843588896969ef2d6add26c1c67ff44a27f7de
Reviewed-on: https://bazel-review.googlesource.com/#/c/2695/
MOS_MIGRATED_REVID=112678600