Return opaque value for native.rule() on a select() statement.
This is imperfect in many ways:

 1) the value is opaque, and not a BUILD value, so it cannot be used
 in rule arguments

 2) its representation has a pointer address, so it breaks hermeticity.

Despite its problem, we return this value because otherwise
native.rules() fails if there is any rule using a select() in the
BUILD file.

A future solution would be to convert BuildType.SelectorList back to
syntax.SelectorList. To do so, we would have to

1) recurse into the Selector contents of SelectorList, so those values
are run through skylarkifyValue too

2) get the right Class<?> value. We could probably get at that by looking at
   ((SelectorList)val).getSelectors().first().getEntries().first().getClass().

--
MOS_MIGRATED_REVID=113051612