Remove the genfilesPath parameter from FdoSupport.prepareToBuild.
This makes the fdo support independent of the configuration, and therefore
safe to cache.

The genfilesPath was only used in AutoFDO builds, which now use the full
exec path to look up profiling data. The profile already refers to imported
files through their full exec paths, so this approach is more consistent.
Furthermore, if we ever have a source file with the same root-relative path as
a generated file, these would conflict in the FdoSupport map, potentially
leading to errors in the build. This may not be possible right now since such
files would conflict in the label namespace and such conflicts are always
resolved towards the generated file.

--
MOS_MIGRATED_REVID=114204994