Problem: The --help option is currently broken when allow residue is false as the --help option cannot be parsed and results in exit(2) after emitting the unhelpful message:
    Error parsing command line: Unrecognized option: --help
    Try --help.

In this case, --help is never reached, and so cannot be used.

Solution:
In the proposed code, if any argument is --help, the --help message will be preferred over emitting parsing errors. The process will now exit(0) prior to parsing any other arguments if --help is passed.

--
MOS_MIGRATED_REVID=115495600