Improve handling of exit routines to ensure we respect the exit status.
As part of this, change atexit to receive the name of a function instead of
just a chunk of code.  This way, we can also control the return value of
each individual atexit function and report it to the user, should these
routines fail when the program apparently succeeded.

--
MOS_MIGRATED_REVID=115575895