Let ActionContextConsumer#getActionContexts() return a Multimap instead of a Map.
This makes it possible to request multiple implementations of the same ActionContext to be available via Executor#getContext().

Currently, specialized SpawnActionContexts like the sandbox or the worker strategy that might have to do a fallback each instantiate their own private copy of e.g. the StandaloneSpawnStrategy. With this change, they can instead get a global instance from the Executor.

--
MOS_MIGRATED_REVID=115705811