Cleanup AbstractQueueVisitorTest, simpleCounter in particular
Second in a sequence of CLs to deflake #simpleCounter. The condition,
commented out in a previous CL, was checking a property that wasn't
deterministic. The maxRunningConcurrently could be either one or two
depending on how quickly the AbstractQueueVisitor scheduled the
runnables.

Also added documentation to a confusing constructor call and removed
dead code.

--
MOS_MIGRATED_REVID=116995523